feat: add Hello component and update routing structure
